Refactor project structure and fix module paths
