+++ index.tsx
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 import { I18nextProvider } from 'react-i18next';
 import App from './src/App';
-import { UserProvider } from './context/UserContext';
+import { UserProvider } from './src/context/UserContext';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) {
