+++ src/journal/JournalDetailModal.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { JournalEntry } from '../../types';
+import { JournalEntry } from '../types';
 import { X } from 'lucide-react';
 
 interface JournalDetailModalProps {
