+++ src/pages/EmotionalJournalPage.tsx
@@ -1,12 +1,12 @@
 
 import React, { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useUser } from '../../context/UserContext';
-import JournalEntryFlow from '../components/journal/JournalEntryFlow';
+import { useUser } from '../context/UserContext';
+import JournalEntryFlow from '../journal/JournalEntryFlow';
 import { JournalEntry } from '../types';
 import { Sparkles, PlusCircle } from 'lucide-react';
-import JournalHistoryList from '../components/journal/JournalHistoryList';
-import EditJournalEntryModal from '../components/journal/EditJournalEntryModal';
+import JournalHistoryList from '../journal/JournalHistoryList';
+import EditJournalEntryModal from '../journal/EditJournalEntryModal';
 
 const EmotionalJournalPage: React.FC = () => {
     const { t } = useTranslation();
