+++ src/pages/WisdomDropsPage.tsx
@@ -1,7 +1,7 @@
 
 import React, { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useUser } from '../../context/UserContext';
+import { useUser } from '../context/UserContext';
 import { Droplets } from 'lucide-react';
 import { WisdomDrop } from '../types';
 import ReflectiveStoryInfographic from '../components/wisdom/ReflectiveStoryInfographic';
